Backend/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
paho-mqtt==2.1.0
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2

